Moved custom hooks to separate folder
